requirements.txt
flet
firebase-admin
Pillow
